+++ app/src/main/java/com/example/weatherapp/data/db/unitlocalized/UnitSpecificCurrentWeatherEntry.kt
@@ -1,12 +1,12 @@
 package com.example.weatherapp.data.db.unitlocalized
 
 interface UnitSpecificCurrentWeatherEntry {
-    val temperature: Int
+    val temperature: Double
     val conditionText: String
     val conditionIconUrl: String
     val windSpeed: Double
     val windDirection: String
     val precipitaionVolume: Double
-    val feelsLikeTemperature: Int
+    val feelsLikeTemperature: Double
     val visibilityDistance: Double
 }
